add transition and slide trasition animation to AboutFragment
